chore: move examples outside
