+++ examples/characters/characters.go
