+++ examples/example.go
@@ -4,10 +4,11 @@ import (
 	"log"
 	"time"
 
+	"github.com/aerogear/graphql-link/examples/characters"
+	"github.com/aerogear/graphql-link/examples/shows"
+	starwars_starship "github.com/aerogear/graphql-link/examples/starwars-starship"
+	"github.com/aerogear/graphql-link/examples/starwars_characters"
 	"github.com/aerogear/graphql-link/internal/gateway"
-	"github.com/aerogear/graphql-link/internal/gateway/examples/characters"
-	"github.com/aerogear/graphql-link/internal/gateway/examples/shows"
-	"github.com/aerogear/graphql-link/internal/gateway/examples/starwars_characters"
 )
 
 func main() {
@@ -27,13 +28,21 @@ func main() {
 	}
 	defer showsServer.Close()
 
-	log.Println("===== Starwars =====")
+	log.Println("===== Starwars Characters =====")
 	starwars_charactersEngine := starwars_characters.New()
 	starwars_charactersServer, err := gateway.StartServer("0.0.0.0", 8083, starwars_charactersEngine, gateway.SimpleLog)
 	if err != nil {
 		panic(err)
 	}
 	defer starwars_charactersServer.Close()
+
+	log.Println("===== Starwars StarShip =====")
+	starwars_starshipEngine := starwars_starship.New()
+	starwars_starshipServer, err := gateway.StartServer("0.0.0.0", 8084, starwars_starshipEngine, gateway.SimpleLog)
+	if err != nil {
+		panic(err)
+	}
+	defer starwars_starshipServer.Close()
 
 	for {
 		time.Sleep(time.Hour)
