+++ examples/starwars_characters/schema.go
