+++ examples/starwars_characters/search_result.go
